BueSol_NET - My first post onto github
Written in VB.Net 2015.  A set of wrapper modules and an example program
for use with the Blue Soleil SDK to demonstrate how to interface with
each Bluetooth profile.  Most of the profiles are already supported.
Only tested under x86 compiling due to heavy use of function pointers
and data-type conversions.  All marshalling is done by hand in code,
building/parsing structures out of byte arrays.